+++ esp32_cloud_communication/lora.h
@@ -4,7 +4,7 @@
 #include <Arduino.h>
 
 void lora_setup();
-void lora_transmit(uint8_t * sending_info, int n);
-void lora_receive(uint8_t * receiving_info, int n);
+void lora_transmit(uint16_t * sending_info, int n);
+void lora_receive(uint16_t * receiving_info, int n);
 
 #endif
